+++ src/rics/translation/fetching/__init__.py
@@ -1,13 +1,15 @@
 """Translation using external sources."""
 
-from rics.translation.fetching._fetcher import Fetcher
+from rics.translation.fetching._abstract_fetcher import AbstractFetcher
 from rics.translation.fetching._memory_fetcher import MemoryFetcher
+from rics.translation.fetching._multi_fetcher import MultiFetcher
 from rics.translation.fetching._pandas_fetcher import PandasFetcher
 from rics.translation.fetching._sql_fetcher import SqlFetcher
 
 __all__ = [
-    "Fetcher",
+    "AbstractFetcher",
     "MemoryFetcher",
+    "MultiFetcher",
     "PandasFetcher",
     "SqlFetcher",
 ]